% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_plots_libraryhits.R
\name{plots_libraryhit}
\alias{plots_libraryhit}
\title{Produce Plots from Library Hits}
\usage{
plots_libraryhit(path)
}
\arguments{
\item{path}{Path to the downloaded table of "View all Library Hits" from molecular networking (classic), library search workflows, or feature-based molecular networking in GNPS.}
}
\description{
This function works to generate plots which summarize the 4 different types of tags, including the GNPS Tags, NCBI taxonomy tags, UBERON tags, and lifestyle tags contribtued by users.
}
\examples{
plots_libraryhit(path = "METABOLOMICS-SNETS-V2-f1a1f3a6-view_all_annotations_DB-main.tsv")
}
\author{
Alan K. Jarmusch 2020
}
