% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_plots_tagtemplate.R
\name{plots_tagtemplate}
\alias{plots_tagtemplate}
\title{Produce Plots from Tag Template}
\usage{
plots_tagtemplate(path)
}
\arguments{
\item{path}{Path to the downloaded GNPS tag template (Google Sheets).}
}
\description{
This function works to generate plots from user-completed tag template files (providing the most complete information about the data) which summarize the 4 different types of tags, including the GNPS Tags, NCBI taxonomy tags, UBERON tags, and lifestyle tags contribtued by users.
}
\examples{
plots_tagtemplate(path = "TEST GNPS Tag - Batch Tag Template.tsv")
}
\author{
Alan K. Jarmusch 2020
}
