% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_Lifestyle_libraryhits.R
\name{summary_Lifestyle_libraryhits}
\alias{summary_Lifestyle_libraryhits}
\title{Summary of GNPS Lifestyle Tags - Library Annotations}
\usage{
summary_Lifestyle_libraryhits()
}
\arguments{
\item{library_annotations}{Path to the downloaded table of "View all Library Hits" from molecular networking (classic) or library search workflows in GNPS.}
}
\description{
This function works to generate a summary of the GNPS Lifestyle tags associated with annotated chemicals resulting from spectral library matching in GNPS. GNPS Lifestyle tags are associated with planar InChIKeys.
}
\author{
Alan K. Jarmusch 2020
}
