% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_Lifestyle_tagtemplate.R
\name{summary_Lifestyle_tagtemplate}
\alias{summary_Lifestyle_tagtemplate}
\title{Summary of GNPS Lifestyle Tags - Tag Template}
\usage{
summary_Lifestyle_tagtemplate()
}
\arguments{
\item{tag_template}{Path to the downloaded GNPS tag template (Google Sheets).}
}
\description{
This function works to generate a summary of the GNPS Lifestyle tags associated with annotated chemicals resulting from spectral library matching in GNPS. Users can choose to add new tags using the GNPS tag template. The GNPS tag template can be downloaded and used to produce the summary plots. The GNPS Lifestyle tags are associated with planar InChIKeys.
}
\author{
Alan K. Jarmusch 2020
}
