% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_NCBI_libraryhits.R
\name{summary_NCBI_libraryhits}
\alias{summary_NCBI_libraryhits}
\title{Summary of NCBI Taxonomy Tags in GNPS - Library Annotations}
\usage{
summary_NCBI_libraryhits()
}
\arguments{
\item{library_annotations}{Path to the downloaded table of "View all Library Hits" from molecular networking (classic) or library search workflows in GNPS.}
}
\description{
This function works to generate a summary of the NCBI taxonomy tags associated with annotated chemicals resulting from spectral library matching in GNPS. NCBI taxonomy tags are associated with planar InChIKeys.
}
\author{
Alan K. Jarmusch 2020
}
