% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_UBERON_tagtemplate.R
\name{summary_UBERON_tagtemplate}
\alias{summary_UBERON_tagtemplate}
\title{Summary of UBERON Tags in GNPS - Tag Template}
\usage{
summary_UBERON_tagtemplate()
}
\arguments{
\item{tag_template}{Path to the downloaded GNPS tag template (Google Sheets).}
}
\description{
This function works to generate a summary of the UBERON tags associated with annotated chemicals resulting from spectral library matching in GNPS. Users can choose to add new tags using the GNPS tag template. The GNPS tag template can be downloaded and used to produce the summary plots. UBERON tags are associated with planar InChIKeys.
}
\author{
Alan K. Jarmusch 2020
}
